langchain-rfp/requirements.txt
PyPDF2
python-docx
python-pptx
pandas
tiktoken
langchain-openai
langchain-core